Extract chat discussion flow
