+++ service-worker.js
@@ -74,6 +74,7 @@ const APP_SHELL = [
   '/js/chat-attestation.js',
   '/js/chat-actions.js',
   '/js/chat-panel.js',
+  '/js/chat-discussion.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
