+++ tests/test-correctness-phase2.js
@@ -130,6 +130,7 @@ const pwaAppShellAssets = [
   '/js/chat-attestation.js',
   '/js/chat-actions.js',
   '/js/chat-panel.js',
+  '/js/chat-discussion.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
