+++ tests/test-custom-lens.js
@@ -272,6 +272,7 @@ unsub();
 // ─── 9. Wiring: chat.js main send ───
 console.log('\n9. chat.js wiring');
 const chatSrc = read('js/chat.js');
+const chatDiscussionSrc = read('js/chat-discussion.js');
 const chatPanelSrc = read('js/chat-panel.js');
 assert("imports hasLens from './lens.js'", chatSrc.includes("from './lens.js'"));
 assert('imports queryLens', chatSrc.includes('queryLens'));
@@ -279,7 +280,7 @@ assert('imports injectLensChunks', chatSrc.includes('injectLensChunks'));
 assert('chat-panel imports updateLensIndicator', chatPanelSrc.includes('updateLensIndicator'));
 assert('main send calls hasLens()', chatSrc.includes('if (hasLens())'));
 assert('main send calls queryLensMulti with user text', /await queryLensMulti\(text,/.test(chatSrc));
-assert('multi-persona calls queryLensMulti with msgText', /await queryLensMulti\(msgText,/.test(chatSrc));
+assert('multi-persona calls queryLensMulti with msgText', /await queryLensMulti\(msgText,/.test(chatDiscussionSrc));
 assert('openChatPanel calls updateLensIndicator', chatPanelSrc.includes('updateLensIndicator()'));
 
 // ─── 10. Wiring: focus-card.js lens integration ───
